pyproject.toml
[project]
name = "dental-management"
version = "0.1.0"
description = "A Django-based dental management system"
requires-python = ">=3.12"
dependencies = [
    "boto3==1.38.23",
    "cachetools==5.5.2",
    "celery==5.5.2",
    "celery-types==0.23.0",
    "colorlog==6.9.0",
    "dj_database_url==2.3.0",
    "Django==5.2.1",
    "django-filter==25.1",
    "django-axes[ipware]==8.0.0",
    "django-cors-headers==4.7.0",
    "django-silk==5.3.1",
    "django-split-settings==1.3.2",
    "django-storages==1.14.6",
    "django_redis==5.4.0",
    "djangorestframework==3.16.0",
    "djangorestframework_simplejwt==5.4.0",
    "drf-spectacular==0.28.0",
    "gunicorn==23.0.0",
    "psycopg==3.2.9",
    "pyjwt==2.10.1",
    "python-dotenv==1.1.0",
    "pytz==2025.2",
    "redis==6.1.0",
    "sentry-sdk==2.29.1",
]

[tool.uv]
dev-dependencies = [
    "mypy==1.11.1",
    "pre-commit==3.8.0",
    "pytest==8.3.2",
    "pytest-cov==5.0.0",
    "pytest-django==4.8.0",
    "ruff==0.6.2",
]

[tool.setuptools]
py-modules = []

[tool.ruff]
show-fixes = true
line-length = 90
exclude = [
    "migrations",
    ".venv",
]

[tool.ruff.lint]
select = ["ALL"]
unfixable = ["T201"]
ignore = [
    "E501",   # Line too long  # Will be fixed by black
    "A003",
    "RUF001",
    "ERA001",
    "F841",   # Local variable is assigned to but never used

    "ANN101",
    "ANN102",  # Missing type annotation for `cls` in classmethod
    "ANN401", # Dynamically typed expressions (typing.Any) are disallowed in `{name}`

    "D10",

    "D203",
    "D213",
    "EXE002",

    "PD011",
    "UP040",  # Type alias uses `TypeAlias` annotation instead of the `type` keyword
]

[tool.ruff.lint.isort]
required-imports = ["from __future__ import annotations"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = [
    "F401",  # Module imported but unused
]
"tests/**" = [
    "S101",  # Use of assert detected
]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports_per_module = true
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true
strict = false
implicit_reexport = true
exclude = [
    "venv",
    "migrations",
]

[tool.pytest.ini_options]
minversion = "7.0"
DJANGO_SETTINGS_MODULE = "api.config.settings"
addopts = "--exitfirst -vv --cov --cov-report=html --cov-fail-under=80"
testpaths = [
    "tests",
]
